# openapi/bookepa-api.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a user account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Authenticate a user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh token and issue a new access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "summary": "Revoke a refresh token session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Return the current authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional retry key. Reusing the same key with the same body returns the original response.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Record a bookkeeping transaction",
        "tags": [
          "transactions"
        ]
      },
      "get": {
        "operationId": "TransactionsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "The last item id from the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "businessId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-05-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-05-31T23:59:59.999Z",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "INCOME",
                "EXPENSE"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "COMPLETED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TransactionResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginatedMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List transactions for a business",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get one transaction",
        "tags": [
          "transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update one transaction",
        "tags": [
          "transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Soft-delete one transaction",
        "tags": [
          "transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bookepa API",
    "description": "Production API contract for the Bookepa backend service.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "development"
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "Ada"
          },
          "lastName": {
            "type": "string",
            "example": "Okafor"
          },
          "phone": {
            "type": "string",
            "example": "+2348012345678"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "correct-horse-42"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Long-lived opaque refresh token. Store securely on mobile devices."
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-42"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "revoked"
        ]
      },
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "example": 20
          },
          "nextCursor": {
            "type": "object"
          }
        },
        "required": [
          "limit"
        ]
      },
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "businessId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "COMPLETED"
            ],
            "default": "COMPLETED"
          },
          "transactionMode": {
            "type": "string",
            "enum": [
              "CASH",
              "TRANSFER"
            ]
          },
          "amount": {
            "type": "number",
            "example": 12500
          },
          "currency": {
            "type": "string",
            "example": "NGN"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "vendorId": {
            "type": "string"
          },
          "paymentMethodId": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "transactionDate": {
            "type": "string",
            "example": "2026-05-02T09:00:00.000Z"
          }
        },
        "required": [
          "businessId",
          "type",
          "amount"
        ]
      },
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "businessId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "COMPLETED"
            ]
          },
          "transactionMode": {
            "type": "string",
            "enum": [
              "CASH",
              "TRANSFER"
            ]
          },
          "amount": {
            "type": "string",
            "example": "12500.00"
          },
          "currency": {
            "type": "string",
            "example": "NGN"
          },
          "description": {
            "type": "object"
          },
          "categoryId": {
            "type": "object"
          },
          "customerId": {
            "type": "object"
          },
          "vendorId": {
            "type": "object"
          },
          "paymentMethodId": {
            "type": "object"
          },
          "reference": {
            "type": "object"
          },
          "transactionDate": {
            "format": "date-time",
            "type": "string"
          },
          "createdById": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "businessId",
          "type",
          "status",
          "amount",
          "currency",
          "transactionDate",
          "createdById",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "COMPLETED"
            ]
          },
          "transactionMode": {
            "type": "string",
            "enum": [
              "CASH",
              "TRANSFER"
            ]
          },
          "amount": {
            "type": "number",
            "example": 12500
          },
          "currency": {
            "type": "string",
            "example": "NGN"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "vendorId": {
            "type": "string"
          },
          "paymentMethodId": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "transactionDate": {
            "type": "string",
            "example": "2026-05-02T09:00:00.000Z"
          }
        }
      },
...
    }
  }
}