# src/auth/dto/refresh-token.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, Length } from 'class-validator';

export class RefreshTokenDto {
  @ApiProperty()
  @IsString()
  @Length(32, 512)
  refreshToken!: string;
}

export class LogoutResponseDto {
  @ApiProperty({ example: true })
  revoked!: boolean;
}